requirements.txt
sendgrid
dnspython